Created constant and utils file
