+++ src/lib/constants.ts
@@ -1,0 +1,55 @@
+export const EXPANSION_ITEMS = [
+  {
+    id: "seafarers",
+    label: "Seafarers",
+  },
+  {
+    id: "citiesAndKnights",
+    label: "Cities and Knights",
+  },
+];
+
+export const ALGORITHMS = [
+  {
+    id: "coastal",
+    label: "Coastal",
+  },
+  {
+    id: "thinLandMass",
+    label: "Thin Land Mass",
+  },
+  {
+    id: "largeLandMass",
+    label: "Large Land Mass",
+  },
+  {
+    id: "largeIslands",
+    label: "Large Islands",
+  },
+  {
+    id: "smallIslands",
+    label: "Small Islands",
+  },
+];
+
+export const NUM_OF_PLAYERS = [
+  {
+    id: "4",
+    label: "4 Players",
+  },
+  {
+    id: "6",
+    label: "6 Players",
+  },
+];
+
+export const BOARD_COLOURS: { [key: string]: string } = {
+  W: "#ccb804",
+  S: "#a5c13f",
+  R: "#7f7f7f",
+  T: "#3a7044",
+  B: "#c7490a",
+  D: "#fceac4",
+  G: "#ffcf00",
+  O: "#3daace",
+};
